Extract a generalized StatusDisplay from Task- and TicketStatus displays
Reuse common code for the TicketStatusDisplay and TaskStatusDisplay by
extracting the StatusDisplay in an own component.
